+++ app/src/main/kotlin/org/fossify/filemanager/extensions/String.kt
@@ -1,3 +1,15 @@
 package org.fossify.filemanager.extensions
 
 fun String.isZipFile() = endsWith(".zip", true)
+
+fun String.isPathInHiddenFolder(): Boolean {
+    val parts = split("/")
+    for (i in 1 until parts.size - 1) {
+        val part = parts[i]
+        val isHidden = part.startsWith(".") && part != "." && part != ".." && part.isNotEmpty()
+        if (isHidden) {
+            return true
+        }
+    }
+    return false
+}
